requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
pyyaml
jupyterlab
xgboost
streamlit
pylint
pytest
mlflow